Feito inserir, listar, alterar, pesquisar e excluir (faltando ver pedidos e excluir pedido do cliente excluido)

+++ Pizzaria_Marchiori/src/main/java/Dominio/Cliente.java
@@ -100,7 +100,7 @@ public class Cliente implements Serializable {
         this.rua = rua;
     }
 
-    public float getNumero() {
+    public int getNumero() {
         return numero;
     }
